Added Roach Colony and more Amenities

+++ src/HotelAmenity.java
@@ -1,7 +1,7 @@
 //The abstract Decorator for the Room
 public abstract class HotelAmenity extends Room
 {
-	public abstract double getCost ();
+	public abstract int getCost();
 
-	public abstract boolean isVacant ();
+	public abstract boolean isVacant();
 }
